% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_species_by_feature}
\alias{get_species_by_feature}
\title{Get species by feature}
\usage{
get_species_by_feature(schema.table, where)
}
\arguments{
\item{schema.table}{The SCHEMA.TABLE containing the feature(s).}

\item{where}{The WHERE clause selecting the feature(s) as the area of interest.}
}
\value{
data.frame of species present in the feature(s)
}
\description{
Output a table of species present from one or more spatial feature(s).
}
\details{
Use the \code{species_by_feature} endpoint of the API at
\href{https://api.marinesensitivity.org/__docs__/#/default/get_species_by_feature}{api.marinesensitivity.org}.
}
\examples{
get_species_by_feature(
  schema.table = "raw.mr_eez",
  where        = "mrgid = 8442")
}
\concept{read}
