% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ms_basemap}
\alias{ms_basemap}
\title{Basemap}
\usage{
ms_basemap(base_opacity = 0.5)
}
\arguments{
\item{base_opacity}{numeric between 0 and 1 (default=0.5)}
}
\value{
\link[leaflet]{leaflet} map object with Esri.OceanBasemap
}
\description{
Basemap with Esri Ocean Basemap
}
\examples{
ms_basemap()
}
\concept{viz}
