% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ply_boemrgns_s05}
\alias{ply_boemrgns_s05}
\title{Polygons of BOEM Regions, simplified to 5\%}
\format{
\subsection{\code{ply_boemrgns_s05}}{

A spatial features \pkg{sf} data frame with 4 rows and 6 columns:
\describe{
\item{boemrgn_key}{key for shelf polygon containing region}
\item{boemrgn_name}{name for shelf polygon containing region}
\item{geom}{geometry of polygon as \code{sf::st_geometry()}}
\item{ctr_lon}{centroid longitude}
\item{ctr_lat}{centroid longitude}
\item{area_km2}{area in square kilometers}
}
}
}
\source{
\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas}
}
\usage{
ply_boemrgns_s05
}
\description{
Polygons describing the BOEM Outer Continental Shelf (OCS) Regions
\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas}
Summarized to region name (key): Alaska (AK), Atlantic (ATL), Gulf of Mexico (GOM), Pacific (PAC).
Simplified to 5\% of original vertices for quickly plotting.
}
\concept{data}
\keyword{datasets}
