% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ply_br}
\alias{ply_br}
\title{Polygons of BOEM Regions}
\format{
A spatial features \pkg{sf} data frame with 5 rows and 6 columns:
\describe{
\item{br_key}{key for BOEM Region}
\item{br_name}{name for BOEM Region}
\item{geom}{geometry of polygon as \code{sf::st_geometry()}}
\item{ctr_lon}{centroid longitude}
\item{ctr_lat}{centroid longitude}
\item{area_km2}{area in square kilometers}
}
}
\source{
\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas}
}
\usage{
ply_br
}
\description{
Polygons describing the BOEM Outer Continental Shelf (OCS) Regions (\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas})
Summarized to region name (key): Alaska (AK), Atlantic (ATL), Gulf of Mexico (GOM), Pacific (PAC).
}
\concept{data}
\keyword{datasets}
