% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ply_ecoareas}
\alias{ply_ecoareas}
\title{Polygons of EcoAreas}
\format{
\subsection{\code{ply_ecoareas}}{

A spatial features \pkg{sf} data frame with 26 rows and 8 columns:
\describe{
\item{ecoarea_key}{key for Planning Area}
\item{ecoarea_name}{name for Planning Area}
\item{geom}{geometry of polygon as \code{sf::st_geometry()}}
\item{ctr_lon}{centroid longitude}
\item{ctr_lat}{centroid longitude}
\item{area_km2}{area in square kilometers}
}
}
}
\usage{
ply_ecoareas
}
\description{
BOEM EcoAreas, which represent the intersection of Ecoregions and Planning Areas.
}
\concept{data}
\keyword{datasets}
