% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ply_ecorgns}
\alias{ply_ecorgns}
\title{Polygons of Ecoregions}
\format{
\subsection{\code{ply_ecorgns}}{

A spatial features \pkg{sf} data frame with 26 rows and 8 columns:
\describe{
\item{ecorgn_key}{key for Planning Area}
\item{ecorgn_name}{name for Planning Area}
\item{geom}{geometry of polygon as \code{sf::st_geometry()}}
\item{ctr_lon}{centroid longitude}
\item{ctr_lat}{centroid longitude}
\item{area_km2}{area in square kilometers}
}
}
}
\usage{
ply_ecorgns
}
\description{
BOEM Ecoregions, extended to the Planning Areas.
}
\concept{data}
\keyword{datasets}
