% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ply_pa}
\alias{ply_pa}
\title{Polygons of Planning Areas}
\format{
A spatial features \pkg{sf} data frame with 26 rows and 8 columns:
\describe{
\item{pa_key}{key for Planning Area}
\item{pa_name}{name for Planning Area}
\item{geom}{geometry of polygon as \code{sf::st_geometry()}}
\item{ctr_lon}{centroid longitude}
\item{ctr_lat}{centroid longitude}
\item{area_km2}{area in square kilometers}
}
}
\source{
\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas}
and \href{https://marinecadastre-noaa.hub.arcgis.com/datasets/noaa::outer-continental-shelf-lands-act/about}{MarineCadastre Outer Continental Shelf Lands Act}
}
\usage{
ply_pa
}
\description{
BOEM Planning Areas (\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas})
}
\concept{data}
\keyword{datasets}
