% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ply_pa_s05}
\alias{ply_pa_s05}
\title{Polygons of Planning Areas, simplified to 5\\%}
\format{
A spatial features \pkg{sf} data frame with 27 rows and 8 columns:
\describe{
\item{pa_key}{key for Planning Area}
\item{pa_name}{name for Planning Area}
\item{geom}{geometry of polygon as \code{sf::st_geometry()}}
\item{ctr_lon}{centroid longitude}
\item{ctr_lat}{centroid longitude}
\item{area_km2}{area in square kilometers}
}
}
\source{
\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas}
}
\usage{
ply_pa_s05
}
\description{
BOEM Planning Areas (\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas})
}
\concept{data}
\keyword{datasets}
