% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ply_planareas}
\alias{ply_planareas}
\title{Polygons of Planning Areas}
\format{
\subsection{\code{ply_planareas}}{

A spatial features \pkg{sf} data frame with 26 rows and 8 columns:
\describe{
\item{planarea_key}{key for Planning Area}
\item{planarea_name}{name for Planning Area}
\item{geom}{geometry of polygon as \code{sf::st_geometry()}}
\item{ctr_lon}{centroid longitude}
\item{ctr_lat}{centroid longitude}
\item{area_km2}{area in square kilometers}
}
}
}
\source{
\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas}
}
\usage{
ply_planareas
}
\description{
BOEM Planning Areas from (\href{https://www.arcgis.com/home/item.html?id=576ae15675d747baaec607594fed086e}{BOEM Offshore Oil and Gas Planning Areas}).
}
\concept{data}
\keyword{datasets}
